setting up chatgpt intergeration with tender data
